+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Allow AC "useOwnTemperature" without using Mode.Auto by splitting the settings.
+
 ## 3.0.0-alpha.18 (2024-03-21)
 
 * (theimo1221) Fix WLED turn on/off process